% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_and_server.R
\name{run_visagejr}
\alias{run_visagejr}
\title{Launch the ViSAGEjr R Shiny app}
\usage{
run_visagejr()
}
\value{
A Shiny app object that launches the ViSAGEjr application.
}
\description{
The\code{run_visagejr()} function launches the ViSAGEjr R Shiny app.
The app is designed to simulate and visualize multi-generation selection of breeding populations.
}
